🚧 SDE prey spin

+++ SDE/particle/prey-pradator.py
@@ -15,13 +15,13 @@ x = 2 * np.random.rand(N, D)
 # z = np.random.rand(N2, D)
 z = np.array([
     [-0.1, 0.5],
-    # [2.1, 0.7],
+    [2.1, 0.7],
     # [-0.1, 0.3],
 ])
 N2 = len(z)
 
-T = 41
-frakit = FrameKit(T)
+T = 21
+frakit = FrameKit(T, 6)
 dt = frakit.dt
 
 # fig, ax = plt.subplots()
@@ -32,8 +32,9 @@ sp = ShootPlot()
 camera = Camera(sp.fig)
 
 
-c = 2
+c = 100
 VIDEO_MODE = True
+DRY_RUN = False
 
 
 def prey_social(r):
@@ -62,6 +63,9 @@ def predator_prey(r):
     return - c * hyper_tang(r, 1, 1)
 
 
+z_trace = [z[0]]
+v_trace = []
+t_series = []
 for f in progressbar(range(frakit.frames)):
     t = f * dt
 
@@ -85,27 +89,47 @@ for f in progressbar(range(frakit.frames)):
         vz = np.nansum(predator_social(zz_norm) / zz_norm * zz, 1) / N2 + \
             np.nansum(predator_prey(zx_norm) / zx_norm * zx, 1) / N
 
-    if VIDEO_MODE:
-        sp.ax.scatter(*(x.T), color='black')
-        sp.ax.scatter(*(z.T), color='purple')
-        sp.text('t=%.2f' % t)
-        camera.snap()
-    else:
-        for s in range(T):
-            if abs(t - s) < dt / 2:
-                sp.quiver(x, vx)
-                sp.quiver(z, vz, color='blue')
-                sp.ax.axis('off')
-                sp.ax.set_aspect('equal', 'box')
-                sp.text('t=%.2f' % t)
-                sp.fig.savefig(
-                    f'particle/prey/tanh-two-predator.c={c}.t={t}.pdf')
-                sp.ax.cla()
+    if not DRY_RUN:
+        if VIDEO_MODE:
+            # sp.ax.scatter(*(x.T), color='black')
+            # sp.ax.scatter(*(z.T), color='red')
+            sp.scatter(x, color='black')
+            sp.scatter(z, color='red')
+            sp.text('t=%.2f' % t)
+            camera.snap()
+        else:
+            for s in range(T):
+                if abs(t - s) < dt / 2:
+                    sp.quiver(x, vx)
+                    sp.quiver(z, vz, color='blue')
+                    sp.ax.axis('off')
+                    sp.ax.set_aspect('equal', 'box')
+                    sp.text('t=%.2f' % t)
+                    sp.fig.savefig(
+                        f'particle/prey/tanh-two-predator.c={c}.t={t}.pdf')
+                    sp.ax.cla()
+
+    v_trace.append(vz[0])
 
     x += vx * dt
     z += vz * dt
+    t_series.append(t)
+    z_trace.append(z[0])
 
 if VIDEO_MODE:
     animation = camera.animate()
     animation.save(
-        f'particle/prey/one-prey-pradator{c}.mp4', fps=frakit.FPS, dpi=200)
+        f'particle/prey/two-prey-pradator{c}.mp4', fps=frakit.FPS, dpi=200)
+
+if DRY_RUN:
+    z_trace = np.array(z_trace)
+    v_trace = np.array(v_trace)
+    plt.figure()
+    # plt.plot(t, z_trace[:, 0], label='horizontal')
+    # plt.plot(t, z_trace[:, 1], label='vertical')
+    plt.plot(t_series, v_trace[:, 0], label='horizontal')
+    plt.plot(t_series, v_trace[:, 1], label='vertical')
+    plt.ylabel('v')
+    plt.xlabel('t')
+    plt.legend()
+    plt.show()
